+++ scripts/mod_00_x=Omega_E_v=E_N.py
@@ -13,18 +13,12 @@ from systems.Mod00 import Mod00
 # all parameters
 params = {
     'looper': {
+        'show_progress': True,
         'X': {
-            'var': 'gs',
+            'var': 'Omegas',
             'idx': 0,
-            'min': 0.000,
-            'max': 0.005,
-            'dim': 51
-        },
-        'Y': {
-            'var': 'gs',
-            'idx': 1,
-            'min': 0.000,
-            'max': 0.005,
+            'min': 1.0,
+            'max': 2.0,
             'dim': 51
         }
     },
@@ -49,29 +43,23 @@ params = {
         'gs': [0.005, 0.005],
         'kappa': 0.15,
         'n_ths': [0, 0],
-        'Omegas': [1.0, 1.0],
+        'Omegas': [1.0, 1.14],
         'omegas': [1.0, 1.0],
         'Vs': [10.0, 1.0]
     },
     'plotter': {
-        'type': 'contourf',
-        'palette': 'Blues',
-        'bins': 11,
-        'x_label': '$g_{0}$',
+        'type': 'line',
+        'x_label': '$\\Omega_{E} / \\omega_{0}$',
         'x_bound': 'both',
-        'x_ticks': [0.000, 0.001, 0.002, 0.003, 0.004, 0.005],
-        'y_label': '$g_{1}$',
-        'y_bound': 'both',
-        'y_ticks': [0.000, 0.001, 0.002, 0.003, 0.004, 0.005],
-        'show_cbar': True,
-        'cbar_position': 'top',
-        'cbar_title': '$E_{N}$',
-        'cbar_ticks': [0.0, 0.05, 0.1, 0.15, 0.2]
+        'x_ticks': [1.0, 1.2, 1.4, 1.6, 1.8, 2.0],
+        'v_label': '$E_{N}$',
+        'v_bound': 'both',
+        'v_ticks': [0.0, 0.2, 0.4, 0.6]
     }
 }
 
 # get average entanglement
-looper = wrap_looper(Mod00, params, 'measure_average', 'XYLooper', 'H:/Workspace/VSCode/Python/entan_oem_mod/data/mod_00/E_N_Omega_E_1.0_Omega_V_1.0', True)
+looper = wrap_looper(Mod00, params, 'measure_average', 'XLooper', 'H:/Workspace/VSCode/Python/entan_oem_mod/data/mod_00/E_N_g_0_0.005_g_1_0.005_Omega_V_1.14', True)
 
 # calculate thresholds
 print(looper.get_thresholds())