+++ scripts/prl_00_x=t_v=E_N.py
@@ -1,5 +1,4 @@
 # dependencies
-import numpy as np
 import os 
 import sys