+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## 2021/05/26 - 00 - Average Entanglement
+> Toolbox version 0.7.3
+* Added scripts for calculation of average entanglement for `Mod00`.
+* Minor fixes to scripts to implement wrappers.
+* Updated `README`.
+
 ## 2021/05/21 - 00 - Added RHC Scripts
 > Toolbox version 0.7.2
 * Added script for RHC counts for `Mod00`.